+++ test/integration.test.js
@@ -17,13 +17,13 @@ test('ST → ST', function(t) {
     let d1 = spacetime([2010, 11, 5, 5, 0], tz);
     let str = d1.format('nice');
     let isDST = d1.timezone().current.isDST;
-    t.equal(str, 'December 5th, 5:00am', 'init-time');
+    t.equal(str, 'Dec 5th, 5:00am', 'init-time');
     t.equal(isDST, false, 'init-dst-off ' + tz);
     //move it to a new time in ST
     d1 = d1.add(3, 'days');
     str = d1.format('nice');
     isDST = d1.timezone().current.isDST;
-    t.equal(str, 'December 8th, 5:00am', 'new-date');
+    t.equal(str, 'Dec 8th, 5:00am', 'new-date');
     t.equal(isDST, false, 'still-dst-off');
   });
   t.end();
@@ -35,13 +35,13 @@ test('ST → DST', function(t) {
     let d1 = spacetime([2010, 11, 5, 10, 0], tz);
     let str = d1.format('nice');
     let isDST = d1.timezone().current.isDST;
-    t.equal(str, 'December 5th, 10:00am', 'init-time');
+    t.equal(str, 'Dec 5th, 10:00am', 'init-time');
     t.equal(isDST, false, 'init-dst-off ' + tz);
     //move it to a new time in DST - April 5th
     d1 = d1.add(4, 'months');
     str = d1.format('nice');
     isDST = d1.timezone().current.isDST;
-    t.equal(str, 'April 5th, 10:00am', 'new-date');
+    t.equal(str, 'Apr 5th, 10:00am', 'new-date');
     t.equal(isDST, true, 'dst-now-on');
   });
   t.end();
@@ -53,13 +53,13 @@ test('DST → DST', function(t) {
     let d1 = spacetime([2010, 5, 5, 15, 0], tz);
     let str = d1.format('nice');
     let isDST = d1.timezone().current.isDST;
-    t.equal(str, 'June 5th, 3:00pm', 'init-time');
+    t.equal(str, 'Jun 5th, 3:00pm', 'init-time');
     t.equal(isDST, true, 'init-dst-on ' + tz);
     //move it to a new time in DST - July 5th
     d1 = d1.add(1, 'months');
     str = d1.format('nice');
     isDST = d1.timezone().current.isDST;
-    t.equal(str, 'July 5th, 3:00pm', 'new-date');
+    t.equal(str, 'Jul 5th, 3:00pm', 'new-date');
     t.equal(isDST, true, 'dst-still-on');
   });
   t.end();
@@ -71,13 +71,13 @@ test('DST → ST', function(t) {
     let d1 = spacetime([2010, 5, 5, 20, 0], tz);
     let str = d1.format('nice');
     let isDST = d1.timezone().current.isDST;
-    t.equal(str, 'June 5th, 8:00pm', 'init-time');
+    t.equal(str, 'Jun 5th, 8:00pm', 'init-time');
     t.equal(isDST, true, 'init-dst-on ' + tz);
     //move it to a new time in ST - December 5th
     d1 = d1.add(6, 'months');
     str = d1.format('nice');
     isDST = d1.timezone().current.isDST;
-    t.equal(str, 'December 5th, 8:00pm', 'new-date');
+    t.equal(str, 'Dec 5th, 8:00pm', 'new-date');
     t.equal(isDST, false, 'dst-now-off');
   });
   t.end();