+++ changelog.md
@@ -69,8 +69,8 @@
 
 # v5.0.0 🚨  -breaking
 * 🚨 all methods now immutable by default.🚨
- - - existing code will now need to be `s = s.add(1,'day')`
- - - - (instead of `s.add(1,'day')`)
+ -- existing code will now need to be `s = s.add(1,'day')`
+ --- (instead of `s.add(1,'day')`)
 * move unix-formating out of `.format()` into `.unixFmt()`
 * support new `{hour} oclock`-style format templates
 * change default `.format()` response to **yyyy-mm-dd**